+++ packages/terra-avatar/src/variants/Facility.jsx
@@ -28,7 +28,7 @@ const propTypes = {
    */
   image: PropTypes.string,
   /**
-   * Whether to hide avatar from the accessiblity tree.
+   * Whether to hide avatar from the accessibility tree.
    */
   isAriaHidden: PropTypes.bool,
   /**
@@ -79,7 +79,7 @@ class Facility extends React.Component {
       facilityContent = generateImagePlaceholder(alt, isAriaHidden, AVATAR_VARIANTS.FACILITY);
     }
     const attributes = { ...customProps };
-    const customStyles = size ? Object.assign({ fontSize: size }, attributes.style) : attributes.style;
+    const customStyles = size ? ({ fontSize: size, ...attributes.style }) : attributes.style;
     const facilityClassNames = cx([
       'avatar',
       setColor(alt, color, hashValue),
@@ -88,11 +88,13 @@ class Facility extends React.Component {
       attributes.className,
     ]);
 
+    /* eslint-disable react/forbid-dom-props */
     return (
       <div {...attributes} className={facilityClassNames} style={customStyles}>
         {facilityContent}
       </div>
     );
+    /* eslint-enable react/forbid-dom-props */
   }
 }