+++ packages/terra-form-checkbox/src/CheckboxUtil.js
@@ -5,7 +5,7 @@ const CheckboxUtil = (() => {
       'ontouchstart' in window
       // eslint-disable-next-line no-undef
       || (window.DocumentTouch && document instanceof DocumentTouch)
-      || navigator.maxTouchPoints > 0
+      || navigator.maxTouchPoints > 0 // eslint-disable-line compat/compat
       || navigator.msMaxTouchPoints > 0
     );