+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/DisabledPlaceholderTextarea.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Textarea from '../../../Textarea';
 
-const textarea = () => <Textarea disabled placeholder="placeholder" />;
+const textarea = () => <Textarea disabled placeholder="placeholder" ariaLabel="label" />;
 
 export default textarea;