+++ packages/terra-icon/scripts/src/migrate-svg/optimizeSvg.js
@@ -4,7 +4,7 @@ import optimizeConfig from './optimizeConfig';
 /* eslint-enable import/no-extraneous-dependencies */
 
 const optimizeSvg = csvObject => new Promise((resolve, reject) => {
-  const objectCsv = Object.assign({}, csvObject);
+  const objectCsv = { ...csvObject };
   const config = optimizeConfig(objectCsv);
   const svgo = new SVGO(config);